[ style ] 컴포넌트 안으로 넣기
