+++ src/@components/profile/index.tsx
@@ -1,6 +1,8 @@
+import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
-import { SleeperAccountIc } from "../../assets";
+import { ProfileEditBtnIc, SleeperAccountIc } from "../../assets";
 import { UserProfileType } from "../../type/profile";
+import BackButton from "../@common/backButton";
 import ProducerProfileImage from "./producerProfileImage";
 import ProfileCategory from "./profileCategory";
 import ProfileDescription from "./profileDescription";
@@ -9,14 +11,26 @@ import VocalProfileImage from "./vocalProfileImage";
 
 interface ProfileProps {
   userType: string | undefined;
+  userSelf: boolean | undefined;
   userProfile: UserProfileType | undefined;
 }
 
 export default function Profile(props: ProfileProps) {
-  const { userType, userProfile } = props;
+  const { userType, userSelf, userProfile } = props;
+  const navigate = useNavigate();
+
+  function handleMoveProfileEditPage() {
+    userType === "vocal"
+      ? navigate(`/profile-edit/vocal/${userProfile?.userId}`)
+      : navigate(`/profile-edit/producer/${userProfile?.userId}`);
+  }
 
   return (
     <ProfileWrapper>
+      <BackButtonWrapper>
+        <BackButton />
+        {userSelf && <ProfileEditBtnIcon onClick={handleMoveProfileEditPage} />}
+      </BackButtonWrapper>
       {userType === "vocal" && <VocalProfileImage />}
       {userType === "producer" && <ProducerProfileImage />}
       <Title>
@@ -32,6 +46,21 @@ export default function Profile(props: ProfileProps) {
     </ProfileWrapper>
   );
 }
+
+const BackButtonWrapper = styled.div`
+  margin: 6rem 0 6rem 0;
+  display: flex;
+  padding: 0 8rem;
+  width: 100%;
+  justify-content: space-between;
+`;
+
+const ProfileEditBtnIcon = styled(ProfileEditBtnIc)`
+  width: 16.6rem;
+  margin-left: 16.9rem;
+
+  cursor: pointer;
+`;
 
 const Title = styled.header`
   margin-top: 2.1rem;
